4.18-4.21

+++ controllers/blogs.js
@@ -1,7 +1,16 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const { error } = require('../utils/logger')
+// const { error } = require('../utils/logger')
+
+const getTokenFrom = request => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.startsWith('Bearer ')) {
+        return authorization.replace('Bearer ', '')
+    }
+    return null
+}
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -12,13 +21,19 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-    const {title, author, url, likes, userId} = request.body
+    const { title, author, url, likes, userId } = request.body
 
     if (!title || !url) {
         return response.status(400).json({ error: "title or url missing" })
     }
 
-    let user = await User.findById(userId)
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
     if (!user) {
         return response.status(400).json({ error: 'invalid user id' })
     }
@@ -37,14 +52,26 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    const id = request.params.id
-    const deleteBlog = await Blog.findByIdAndDelete(id)
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
 
-    if (deleteBlog) {
-        response.status(204).end()
-    } else {
-        response.status(404).json({ error: 'blog not found' })
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
     }
+
+    const id = request.params.id
+    const blog = await Blog.findById(id)
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (blog.user.toString() !== decodedToken.id) {
+        return response.status(403).json({ error: 'unauthorized action' })
+    }
+
+    await Blog.findByIdAndDelete(id)
+    response.status(204).end()
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
